third_party/WebKit/LayoutTests/fast/forms/associatedFormControls-leak-nodes.html

<!DOCTYPE html>
<html>
<body>
<iframe id="frame"></iframe>

<script src="../../resources/js-test.js"></script>
<script>
if (!window.internals) {
    debug("This test only runs on \"content_shell --dump-render-tree\", as it requires existence of window.internals.");
} else {
    gc();
    var documentsBefore = window.internals.numberOfLiveDocuments();

    var frame = document.getElementById('frame');
    frame.contentDocument.body.innerHTML = '<form></form>';
    document.body.removeChild(frame);
    frame = null;

    gc();
    var documentsAfter = window.internals.numberOfLiveDocuments();

    // -1 is from removing frame itself.
    shouldBe('documentsBefore - 1', 'documentsAfter');
}
</script>
</body>
</html>
